[spark] Fixes python tarslip security concern

+++ basicdataset/src/main/resources/imagenet/extract_imagenet.py
@@ -156,6 +156,20 @@ def build_rec_process(img_dir, train=False, num_thread=1):
     os.remove(lst_path)
     print('ImageRecord file for ' + prefix + ' has been built!')
 
+def is_within_directory(directory, target):
+    abs_directory = os.path.abspath(directory)
+    abs_target = os.path.abspath(target)
+    prefix = os.path.commonprefix([abs_directory, abs_target])
+    return prefix == abs_directory
+
+def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
+    for member in tar.getmembers():
+        member_path = os.path.join(path, member.name)
+        if not is_within_directory(path, member_path):
+            raise Exception("Attempted Path Traversal in Tar File")
+    tar.extractall(path, members, numeric_owner=numeric_owner)
+
+
 def extract_train(tar_fname, target_dir, with_rec=False, num_thread=1):
     os.makedirs(target_dir)
     with tarfile.open(tar_fname) as tar:
@@ -164,31 +178,16 @@ def extract_train(tar_fname, target_dir, with_rec=False, num_thread=1):
         pbar = tqdm(total=len(tar.getnames()))
         for class_tar in tar:
             pbar.set_description('Extract '+class_tar.name)
-            tar.extract(class_tar, target_dir)
             class_fname = os.path.join(target_dir, class_tar.name)
+            if not is_within_directory(target_dir, class_fname):
+                raise Exception("Attempted Path Traversal in Tar File")
+
+            tar.extract(class_tar, target_dir)
             class_dir = os.path.splitext(class_fname)[0]
             os.mkdir(class_dir)
             with tarfile.open(class_fname) as f:
-                def is_within_directory(directory, target):
-                    
-                    abs_directory = os.path.abspath(directory)
-                    abs_target = os.path.abspath(target)
-                
-                    prefix = os.path.commonprefix([abs_directory, abs_target])
-                    
-                    return prefix == abs_directory
-                
-                def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-                
-                    for member in tar.getmembers():
-                        member_path = os.path.join(path, member.name)
-                        if not is_within_directory(path, member_path):
-                            raise Exception("Attempted Path Traversal in Tar File")
-                
-                    tar.extractall(path, members, numeric_owner=numeric_owner) 
-                    
-                
                 safe_extract(f, class_dir)
+
             os.remove(class_fname)
             pbar.update(1)
         pbar.close()
@@ -199,26 +198,8 @@ def extract_val(tar_fname, target_dir, with_rec=False, num_thread=1):
     os.makedirs(target_dir)
     print('Extracting ' + tar_fname)
     with tarfile.open(tar_fname) as tar:
-        def is_within_directory(directory, target):
-            
-            abs_directory = os.path.abspath(directory)
-            abs_target = os.path.abspath(target)
-        
-            prefix = os.path.commonprefix([abs_directory, abs_target])
-            
-            return prefix == abs_directory
-        
-        def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-        
-            for member in tar.getmembers():
-                member_path = os.path.join(path, member.name)
-                if not is_within_directory(path, member_path):
-                    raise Exception("Attempted Path Traversal in Tar File")
-        
-            tar.extractall(path, members, numeric_owner=numeric_owner) 
-            
-        
         safe_extract(tar, target_dir)
+
     # build rec file before images are moved into subfolders
     if with_rec:
         build_rec_process(target_dir, False, num_thread)
